requirements.txt
autogen
pyautogen
dask[dataframe]
python-dotenv
chromadb
autogen-agentchat[retrievechat]~=0.2